feat: switched to client side data fetching on user page
